BirthdayPostStash/requirements.txt
celery==3.1.25
Django==1.11.7
django-anymail==1.2
django-autoslug==1.9.3
django-celery==3.2.1
django-cors-headers==1.2.0
django-debug-toolbar==1.8
django-environ==0.4.4
django-filter==1.0.4
django-haystack==2.6.1
django-push-notifications==1.5.0
django-redis==4.8.0
django-sql-explorer==1.1.1
djangorestframework==3.6.4
elasticsearch==5.4.0
gunicorn==19.7.1
mixer==5.6.6
mock==2.0.0
psycopg2==2.7.3.1
python-dateutil==2.6.1
python-quickbooks==0.7.3
pytz==2017.2
requests==2.18.4
django-widget-tweaks==1.4.2
ipython
imagekit
Pillow
MySQL-python